Move Java2D functionality into Java2D module.
