+++ openhtmltopdf-java2d/src/main/java/com/openhtmltopdf/java2d/font/AWTFSFont.java
@@ -1,6 +1,6 @@
 /*
  * {{{ header & license
- * Copyright (c) 2007 Wisconsin Court System
+ * Copyright (c) 2006 Wisconsin Court System
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public License
@@ -9,7 +9,7 @@
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public License
@@ -17,20 +17,28 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  * }}}
  */
-package com.openhtmltopdf.swing;
+package com.openhtmltopdf.java2d.font;
 
-import java.awt.font.GlyphVector;
+import java.awt.Font;
+import java.util.List;
 
-import com.openhtmltopdf.extend.FSGlyphVector;
+import com.openhtmltopdf.render.FSFont;
 
-public class AWTFSGlyphVector implements FSGlyphVector {
-    private final GlyphVector _glyphVector;
-    
-    public AWTFSGlyphVector(GlyphVector vector) {
-        _glyphVector = vector;
+public class AWTFSFont implements FSFont {
+    private final List<Font> _fonts;
+    private final float _size;
+
+    public AWTFSFont(List<Font> fonts, float size) {
+        _fonts = fonts;
+        _size = size;
     }
 
-    public GlyphVector getGlyphVector() {
-        return _glyphVector;
+    @Override
+    public float getSize2D() {
+        return _size;
+    }
+
+    public List<Font> getAWTFonts() {
+        return _fonts;
     }
 }
