+++ openhtmltopdf-java2d/src/main/java/com/openhtmltopdf/java2d/image/AWTFSImage.java
@@ -18,10 +18,9 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  * }}}
  */
-package com.openhtmltopdf.swing;
+package com.openhtmltopdf.java2d.image;
 
 import com.openhtmltopdf.extend.FSImage;
-import com.openhtmltopdf.util.ImageUtil;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -56,10 +55,12 @@ public abstract class AWTFSImage implements FSImage {
             this.img = img;
         }
 
+        @Override
         public int getWidth() {
             return img.getWidth(null);
         }
 
+        @Override
         public int getHeight() {
             return img.getHeight(null);
         }
@@ -69,6 +70,7 @@ public abstract class AWTFSImage implements FSImage {
             return img;
         }
 
+        @Override
         public void scale(int width, int height) {
             if (width > 0 || height > 0) {
                 int currentWith = getWidth();
@@ -94,14 +96,17 @@ public abstract class AWTFSImage implements FSImage {
     private static class NullImage extends AWTFSImage {
         private static final BufferedImage EMPTY_IMAGE = ImageUtil.createTransparentImage(1, 1);
 
+        @Override
         public int getWidth() {
             return 0;
         }
 
+        @Override
         public int getHeight() {
             return 0;
         }
 
+        @Override
         public void scale(int width, int height) {
         }
