+++ openhtmltopdf-java2d/src/main/java/com/openhtmltopdf/java2d/image/ImageUtil.java
@@ -17,15 +17,12 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  * }}}
  */
-package com.openhtmltopdf.util;
+package com.openhtmltopdf.java2d.image;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
-import java.util.Base64;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.logging.Level;
-import java.util.regex.Pattern;
 
 /**
  * Static utility methods for working with images. Meant to suggest "best practices" for the most straightforward
@@ -203,29 +200,6 @@ public class ImageUtil {
         return bi;
     }
 
-    private static final Pattern WHITE_SPACE = Pattern.compile("\\s+");
-
-    public static byte[] fromBase64Encoded(String b64encoded) {
-        return Base64.getMimeDecoder().decode(WHITE_SPACE.matcher(b64encoded).replaceAll(""));
-    }
-
-    /**
-     * Get the binary content of an embedded base 64 image.
-     *
-     * @param imageDataUri URI of the embedded image
-     * @return The binary content
-     */
-    public static byte[] getEmbeddedBase64Image(String imageDataUri) {
-        int b64Index = imageDataUri.indexOf("base64,");
-        if (b64Index != -1) {
-            String b64encoded = imageDataUri.substring(b64Index + "base64,".length());
-            return fromBase64Encoded(b64encoded);
-        } else {
-            XRLog.log(Level.SEVERE, LogMessageId.LogMessageId0Param.LOAD_EMBEDDED_DATA_URI_MUST_BE_ENCODED_IN_BASE64);
-        }
-        return null;
-    }
-    
     interface Scaler {
         /**
          * Convenience method that returns a scaled instance of the
