+++ openhtmltopdf-pdfbox/src/main/java/com/openhtmltopdf/pdfboxout/PdfBoxReplacedElementFactory.java
@@ -24,8 +24,8 @@ import com.openhtmltopdf.layout.LayoutContext;
 import com.openhtmltopdf.outputdevice.helper.ExternalResourceType;
 import com.openhtmltopdf.render.BlockBox;
 import com.openhtmltopdf.resource.XMLResource;
+import com.openhtmltopdf.swing.NaiveUserAgent;
 
-import com.openhtmltopdf.util.ImageUtil;
 import org.w3c.dom.Element;
 
 import java.io.ByteArrayInputStream;
@@ -63,7 +63,9 @@ public class PdfBoxReplacedElementFactory implements ReplacedElementFactory {
                 //handle the case of linked svg from img tag
                 boolean isDataImageSvg = false;
                 if (_svgImpl != null && (srcAttr.endsWith(".svg") || (isDataImageSvg = srcAttr.startsWith("data:image/svg+xml;base64,")))) {
-                    XMLResource xml = isDataImageSvg ? XMLResource.load(new ByteArrayInputStream(ImageUtil.getEmbeddedBase64Image(srcAttr))) : uac.getXMLResource(srcAttr, ExternalResourceType.XML_SVG);
+                    XMLResource xml = isDataImageSvg ?
+                         XMLResource.load(new ByteArrayInputStream(NaiveUserAgent.getEmbeddedBase64Image(srcAttr))) : 
+                         uac.getXMLResource(srcAttr, ExternalResourceType.XML_SVG);
 
                     if (xml != null) {
                         Element svg = xml.getDocument().getDocumentElement();
